forum: move cancel tasks into separate files
